Add core layer
