--- src/homeworks-api/src/AspnetRun.Core/Configuration/AspnetRunSettings.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace AspnetRun.Core.Configuration
-{
-    public class AspnetRunSettings
-    {
-        public string ConnectionString { get; set; }
-    }
-}
