--- src/homeworks-api/src/AspnetRun.Core/Entities/Base/Entity.cs
@@ -1,6 +1,0 @@
-﻿namespace AspnetRun.Core.Entities.Base
-{
-    public abstract class Entity : EntityBase<int>
-    {
-    }
-}
