--- src/homeworks-api/src/AspnetRun.Core/Entities/Base/IEntityBase.cs
@@ -1,7 +1,0 @@
-﻿namespace AspnetRun.Core.Entities.Base
-{
-    public interface IEntityBase<TId>
-    {
-        TId Id { get; }
-    }
-}
