--- src/homeworks-api/src/AspnetRun.Core/Exceptions/CoreException.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace AspnetRun.Core.Exceptions
-{
-    public class CoreException : Exception
-    {
-        internal CoreException(string businessMessage)
-            : base(businessMessage)
-        {
-        }
-
-        internal CoreException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
-    }
-}
