--- src/homeworks-api/src/AspnetRun.Core/Interfaces/IAppLogger.cs
@@ -1,8 +1,0 @@
-﻿namespace AspnetRun.Core.Interfaces
-{
-    public interface IAppLogger<T>
-    {
-        void LogInformation(string message, params object[] args);
-        void LogWarning(string message, params object[] args);
-    }
-}
