--- src/homeworks-api/src/AspnetRun.Core/Repositories/Base/IRepository.cs
@@ -1,30 +1,0 @@
-﻿using AspnetRun.Core.Entities.Base;
-using AspnetRun.Core.Specifications.Base;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Core.Repositories.Base
-{
-    public interface IRepository<T> where T : Entity
-    {
-        Task<IReadOnlyList<T>> GetAllAsync();
-        Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>> predicate);
-        Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>> predicate = null,
-                                        Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null,
-                                        string includeString = null,
-                                        bool disableTracking = true);
-        Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>> predicate = null,
-                                        Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null,
-                                        List<Expression<Func<T, object>>> includes = null,
-                                        bool disableTracking = true);
-        Task<IReadOnlyList<T>> GetAsync(ISpecification<T> spec);
-        Task<T> GetByIdAsync(int id);
-        Task<T> AddAsync(T entity);
-        Task UpdateAsync(T entity);
-        Task DeleteAsync(T entity);
-        Task<int> CountAsync(ISpecification<T> spec);
-    }
-}
