--- src/homeworks-api/src/AspnetRun.Core/Repositories/ICategoryRepository.cs
@@ -1,11 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using AspnetRun.Core.Repositories.Base;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Core.Repositories
-{
-    public interface ICategoryRepository : IRepository<Category>
-    {
-        Task<Category> GetCategoryWithProductsAsync(int categoryId);
-    }
-}
