--- src/homeworks-api/src/AspnetRun.Core/Repositories/IProductRepository.cs
@@ -1,14 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using AspnetRun.Core.Repositories.Base;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Core.Repositories
-{
-    public interface IProductRepository : IRepository<Product>
-    {
-        Task<IEnumerable<Product>> GetProductListAsync();
-        Task<IEnumerable<Product>> GetProductByNameAsync(string productName);
-        Task<IEnumerable<Product>> GetProductByCategoryAsync(int categoryId);
-    }
-}
