+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/Configuration/HomeworkSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace EKids.Chatbot.Homeworks.Core.Configuration;
+
+public class HomeworkSettings
+{
+    public required string ConnectionString { get; set; }
+}
