+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/Entities/Base/Entity.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace EKids.Chatbot.Homeworks.Core.Entities.Base;
+
+public abstract class Entity(Guid id) : EntityBase<Guid>(id);
