+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/Entities/Base/IEntityBase.cs
@@ -1,0 +1,6 @@
+﻿namespace EKids.Chatbot.Homeworks.Core.Entities.Base;
+
+public interface IEntityBase<TId>
+{
+    TId Id { get; }
+}
