+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/Exceptions/CoreException.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace EKids.Chatbot.Homeworks.Core.Exceptions;
+
+public class CoreException : Exception
+{
+    internal CoreException(string businessMessage)
+        : base(businessMessage)
+    {
+    }
+
+    internal CoreException(string message, Exception innerException)
+        : base(message, innerException)
+    {
+    }
+}
