+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/GlobalSuppressions.cs
@@ -1,0 +1,8 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Style", "IDE0046:Convert to conditional expression", Justification = "<Pending>", Scope = "member", Target = "~M:EKids.Chatbot.Homeworks.Core.ValueObjects.Base.ValueObject.EqualOperator(EKids.Chatbot.Homeworks.Core.ValueObjects.Base.ValueObject,EKids.Chatbot.Homeworks.Core.ValueObjects.Base.ValueObject)~System.Boolean")]
