+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Core/ValueObjects/Base/ValueObject.cs
@@ -1,0 +1,59 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace EKids.Chatbot.Homeworks.Core.ValueObjects.Base;
+
+// Source: https://docs.microsoft.com/en-us/dotnet/standard/microservices-architecture/microservice-ddd-cqrs-patterns/implement-value-objects
+public abstract class ValueObject
+{
+    protected static bool EqualOperator(ValueObject left, ValueObject right)
+    {
+        if (left is null ^ right is null)
+        {
+            return false;
+        }
+
+        return left?.Equals(right) != false;
+    }
+
+    protected static bool NotEqualOperator(ValueObject left, ValueObject right)
+    {
+        return !EqualOperator(left, right);
+    }
+
+    protected abstract IEnumerable<object> GetAtomicValues();
+
+    public override bool Equals(object? obj)
+    {
+        if (obj == null || obj.GetType() != GetType())
+        {
+            return false;
+        }
+
+        var other = (ValueObject)obj;
+        var thisValues = GetAtomicValues().GetEnumerator();
+        var otherValues = other.GetAtomicValues().GetEnumerator();
+
+        while (thisValues.MoveNext() && otherValues.MoveNext())
+        {
+            if (thisValues.Current is null ^ otherValues.Current is null)
+            {
+                return false;
+            }
+
+            if (thisValues.Current?.Equals(otherValues.Current) == false)
+            {
+                return false;
+            }
+        }
+
+        return !thisValues.MoveNext() && !otherValues.MoveNext();
+    }
+
+    public override int GetHashCode()
+    {
+        return GetAtomicValues()
+            .Select(x => (x?.GetHashCode()) ?? 0)
+            .Aggregate((x, y) => x ^ y);
+    }
+}
